backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.4.5
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
jinja2>=3.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7